Add the SDK changes for multitopics
This change enables functions to listen to multiple topics.  Prior to this
the connector could listen to multiple topics but would read a function's topic
annotation as a single value.  Following changes to the SDK annotations can now
be multiple delimited topics.  The delimiter defaults to a comma and can be
over-ridden using the `topic_delimiter` env var.

Associated changes have been made to the README - to surface the availability of
the variable, and to  `connector-dep.yml` for similar reasons.

Signed-off-by: Richard Gee <richard@technologee.co.uk>